test: update colour_role tests
